move TypeConstraint into src/python/pants/util/objects.py
yet again somehow everything Just Works
