+++ src/python/pants/engine/rules.py
@@ -14,10 +14,9 @@ from types import TypeType
 
 from twitter.common.collections import OrderedSet
 
-from pants.engine.addressable import Exactly
 from pants.engine.selectors import Get, type_or_constraint_repr
 from pants.util.meta import AbstractClass
-from pants.util.objects import datatype
+from pants.util.objects import Exactly, datatype
 
 
 logger = logging.getLogger(__name__)
@@ -87,7 +86,7 @@ class Rule(AbstractClass):
 
 class TaskRule(datatype('TaskRule', ['output_constraint', 'input_selectors', 'input_gets', 'func']), Rule):
   """A Rule that runs a task function when all of its input selectors are satisfied.
-  
+
   TODO: Make input_gets non-optional when more/all rules are using them.
   """
 
@@ -146,7 +145,7 @@ class SingletonRule(datatype('SingletonRule', ['output_constraint', 'value']), R
 
 class RootRule(datatype('RootRule', ['output_constraint']), Rule):
   """Represents a root input to an execution of a rule graph.
-  
+
   Roots act roughly like parameters, in that in some cases the only source of a
   particular type might be when a value is provided as a root subject at the beginning
   of an execution.
