+++ src/python/pants/util/objects.py
@@ -8,7 +8,10 @@ from __future__ import (absolute_import, division, generators, nested_scopes, pr
 import sys
 from collections import OrderedDict, namedtuple
 
+from abc import abstractmethod
+
 from pants.util.memo import memoized
+from pants.util.meta import AbstractClass
 
 
 def datatype(*args, **kwargs):
@@ -79,52 +82,144 @@ class TypeCheckError(TypedDatatypeInstanceConstructionError):
       type_name, formatted_msg, *args, **kwargs)
 
 
-class TypeMatcher(datatype('TypeMatcher', ['types'])):
+class TypeConstraintError(TypeError):
+  """Indicates a :class:`TypeConstraint` violation."""
 
-  class ConstructionError(Exception):
-    pass
 
-  def __new__(cls, types):
-    try:
-      type_set = frozenset(types)
-    except TypeError as e:
-      raise cls.ConstructionError(
-        "'types' argument must be convertible to frozenset: '{}'".format(types))
+class TypeConstraint(AbstractClass):
+  """Represents a type constraint.
 
-    if not type_set:
-      raise cls.ConstructionError("at least one type must be provided to Union")
+  Not intended for direct use; instead, use one of :class:`SuperclassesOf`, :class:`Exact` or
+  :class:`SubclassesOf`.
+  """
 
-    non_type_args = []
-    for ty in type_set:
-      if not isinstance(ty, type):
-        non_type_args.append(ty)
-    if non_type_args:
-      raise cls.ConstructionError(
-        "all elements of 'types' must be a type. non-type args: '{}'"
-        .format(non_type_args))
+  def __init__(self, *types, **kwargs):
+    """Creates a type constraint centered around the given types.
 
-    return super(TypeMatcher, cls).__new__(cls, types=type_set)
+    The type constraint is satisfied as a whole if satisfied for at least one of the given types.
+
+    :param type *types: The focus of this type constraint.
+    :param str description: A description for this constraint if the list of types is too long.
+    """
+    if not types:
+      raise ValueError('Must supply at least one type')
+    if any(not isinstance(t, type) for t in types):
+      raise TypeError('Supplied types must be types. {!r}'.format(types))
+
+    self._types = types
+    self._desc = kwargs.get('description', None)
+
+  @property
+  def types(self):
+    """Return the subject types of this type constraint.
+
+    :type: tuple of type
+    """
+    return self._types
+
+  def satisfied_by(self, obj):
+    """Return `True` if the given object satisfies this type constraint.
+
+    :rtype: bool
+    """
+    return self.satisfied_by_type(type(obj))
+
+  @abstractmethod
+  def satisfied_by_type(self, obj_type):
+    """Return `True` if the given object satisfies this type constraint.
+
+    :rtype: bool
+    """
+
+  def validate_satisfied_by(self, obj):
+    """Return `obj` if it satisfies this type constraint, or raise.
+
+    :raises: `TypeConstraintError` if the given object does not satisfy this
+    type constraint.
+    """
+
+    if self.satisfied_by(obj):
+      return obj
+
+    raise TypeConstraintError(
+      "Value '{}' (with type '{}') did not satisfy type constraint '{!r}'."
+      .format(obj, type(obj).__name__, self))
+
+  def __hash__(self):
+    return hash((type(self), self._types))
+
+  def __eq__(self, other):
+    return type(self) == type(other) and self._types == other._types
+
+  def __ne__(self, other):
+    return not (self == other)
+
+  def __str__(self):
+    if self._desc:
+      constrained_type = '({})'.format(self._desc)
+    else:
+      if len(self._types) == 1:
+        constrained_type = self._types[0].__name__
+      else:
+        constrained_type = '({})'.format(', '.join(t.__name__ for t in self._types))
+    return '{variance_symbol}{constrained_type}'.format(variance_symbol=self._variance_symbol,
+                                                        constrained_type=constrained_type)
+
+  def __repr__(self):
+    if self._desc:
+      constrained_type = self._desc
+    else:
+      constrained_type = ', '.join(t.__name__ for t in self._types)
+    return ('{type_constraint_type}({constrained_type})'
+      .format(type_constraint_type=type(self).__name__,
+                    constrained_type=constrained_type))
+
+
+class SuperclassesOf(TypeConstraint):
+  """Objects of the exact type as well as any super-types are allowed."""
+
+  _variance_symbol = '-'
+
+  def satisfied_by_type(self, obj_type):
+    return any(issubclass(t, obj_type) for t in self._types)
+
+
+class Exactly(TypeConstraint):
+  """Only objects of the exact type are allowed."""
+
+  _variance_symbol = '='
 
   @classmethod
-  def create(cls, decl_val):
-    if isinstance(decl_val, cls):
-      return decl_val
+  def from_type_or_collection(cls, maybe_decl):
+    if isinstance(maybe_decl, cls):
+      return maybe_decl
+    if isinstance(maybe_decl, type):
+      return cls(maybe_decl)
 
-    if isinstance(decl_val, type):
-      return cls(types=[decl_val])
+    try:
+      return cls(*maybe_decl)
+    except TypeError:
+      raise TypeConstraintError(
+        "value '{}' could not be interpreted as an Exactly type constraint, "
+        "a specific type, or a collection of types".format(maybe_decl))
 
-    # Interpret it as an iterable of types, or raise ConstructionError.
-    return cls(types=decl_val)
+  def satisfied_by_type(self, obj_type):
+    return obj_type in self._types
 
-  def matches_value(self, val):
-    """Return whether or not the argument matches the type described by this
-  class."""
-    return type(val) in self.types
+  def graph_str(self):
+    if len(self.types) == 1:
+      return self.types[0].__name__
+    else:
+      return repr(self)
 
-  def compose(self, rhs):
-    """Return a TypeMatcher which matches either this or another TypeMatcher."""
-    all_types = self.types | rhs.types
-    return TypeMatcher(types=all_types)
+
+class SubclassesOf(TypeConstraint):
+  """Objects of the exact type as well as any sub-types are allowed."""
+
+  _variance_symbol = '+'
+
+  def satisfied_by_type(self, obj_type):
+    return issubclass(obj_type, self._types)
 
 
 def typed_datatype(type_name, field_decls):
@@ -140,13 +235,12 @@ def typed_datatype(type_name, field_decls):
     raise TypedDatatypeClassConstructionError(
       type_name,
       "field_decls is not a dict: '{}'".format(field_decls))
-
   if not field_decls:
     raise TypedDatatypeClassConstructionError(
       type_name,
       "no fields were declared")
 
-  # Turn every type declaration into an instance of TypeMatcher, and place the
+  # Turn every type declaration into an instance of Exactly, and place the
   # results in processed_type_decls.
   # TODO: Make this kind of exception pattern (filter for errors then display
   # them all at once) more ergonomic.
@@ -154,11 +248,11 @@ def typed_datatype(type_name, field_decls):
   invalid_type_decls = []
   for name, maybe_decl in field_decls.items():
     try:
-      processed_type_decls[name] = TypeMatcher.create(maybe_decl)
-    except TypeMatcher.ConstructionError as e:
+      processed_type_decls[name] = Exactly.from_type_or_collection(maybe_decl)
+    except TypeConstraintError as e:
       invalid_type_decls.append(
-        "in field '{}': {}".format(name, e))
-
+        "in field '{}': {}"
+        .format(name, e))
   if invalid_type_decls:
     raise TypedDatatypeClassConstructionError(
       type_name,
@@ -168,6 +262,8 @@ def typed_datatype(type_name, field_decls):
 
   datatype_cls = datatype(type_name, list(field_name_set))
 
+  # TODO(cosmicexplorer): Make the repr use the 'description' kwarg in
+  # TypeConstraint?
   class TypedDatatype(datatype_cls):
 
     # We intentionally disallow positional arguments here.
@@ -195,10 +291,12 @@ def typed_datatype(type_name, field_decls):
       type_failure_msgs = []
       for field_name, field_value in kwargs.items():
         field_type = processed_type_decls[field_name]
-        if not field_type.matches_value(field_value):
+        try:
+          field_type.validate_satisfied_by(field_value)
+        except TypeConstraintError as e:
           type_failure_msgs.append(
-            "field '{}' is not an instance of '{}'. type='{}', value='{}'"
-            .format(field_name, field_type, type(field_value), field_value))
+            "field '{}' was invalid: {}"
+            .format(field_name, field_value, e))
       if type_failure_msgs:
         raise TypeCheckError(type_name, '\n'.join(type_failure_msgs))
